backend/requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-llms-openai
